Update Gnash.sh: Add Quality-Set + HW Accel
Bit of a hacky update, may or may not revert in the future.

Let's talk about the smaller change first:  hardware acceleration.  This doesn't really give too much of an increase, but it's not enabled by default so I might as well throw it in.  I also moved the ROM path to the end of the command to make it look more elegant.

Now for the hacky one:  the auto-quality set.  The command looks in the .gnashrc file for the default quality setting (Auto) and changes it to Low.  The hacky part is the fact that, for some reason, Gnash puts the config file in /root/.  This isn't necessarily a *bad* thing, but it's inconsistent with the other emulators and I couldn't find any way to change it.  (The official Gnash user manual does mention that the file is supposed to be in `~/.gnashrc`, so it's entirely possible that something's wrong with `aptInstall` or I didn't set a certain variable or something.)