Gnash.sh: Remove forced 480p, added resolution instructions
This removes the 480p resolution that previous versions of the script installed, and instead added instructions on how to set your own resolution in emulators.cfg.  The primary reason for this is to allow the user to choose their own suitable resolution.  I'd like to simplify the process of configuring resolutions in the future, maybe with a small utility script or something.

Next on the to-do list:  Figure out how to get xinit to run without `sudo` so that Gnash isn't ran as root.   (current snag is denied permissions to access /dev/tty0)