Fixed Sudo quirk, added Xorg as depend, removed auto-quality
Lots of changes here!

First, and the biggest:  **Gnash no longer requires sudo to run.**  Turns out that the main cause of this bug appears to be Stretch rather than anything I or X did.  The fix for this is commented so you can see it in all of it's glory.

A few other minor changes as well, such as the removal of quality being auto-set to 0 and adding Xorg as a depend.