Create pokerth.sh
Tested and working, but without the "usermod -a -G tty pi" command. I executed this on the terminal and I'm unsure if it should be added to configure_pokerth() (as it is now).
It requires root and pi must logout and back in to make the group add take effect.
Playable even without mouse and keyboard, when PS4 controller is used.